Separated the mechanism from the drive train.
