+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftcsim/movement/LineMovement.java
@@ -1,0 +1,4 @@
+package org.firstinspires.ftc.teamcode.ftcsim.movement;
+
+public class LineMovement {
+}
